pyproject.toml
[project]
name = "24-sum-b-4-pe-portfolio-site"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Mike Odnis", email = "airwolf635@gmail.com"},
]
dependencies = [
    "cffi==1.15.0",
    "click==8.0.1",
    "cryptography==42.0.4",
    "flask>=2.2.5",
    "itsdangerous==2.0.1",
    "Jinja2==3.1.4",
    "MarkupSafe",
    "peewee==3.14.10",
    "pycparser==2.21",
    "PyMySQL==1.1.1",
    "python-dotenv==0.17.1",
    "Werkzeug",
    "requests==2.32.2",
    "db-sqlite3",
    "pydantic",
    "pytest",
    "pytest-flask",
    "requests-mock",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
